Move type to correct file
